rendering tests for all components
