--- src/tests/dummy.test.js
@@ -1,7 +1,0 @@
-/* eslint-disable */
-function three(){
-    return 3
-}
-it('numbers', () => {
-  expect(three()).toEqual(3)
-})
